Move to gunicorn/upstart server for graphite-web
For sites that aren't using apache, bringing apache in just for mod_python isn't
ideal. This commit changes the installation procedure so that all python
packages are installed in a virtualenv in /opt/graphite, and installs gunicorn
and a suitable upstart configuration for serving graphite-web with gunicorn on
the specified port.

Using nginx or apache to provide appropriate vhost mappings to the graphite
service is no longer considered the responsibility of this module.