Manually patch graphite-web to support cairocffi
Pycairo is a bit of a nasty module, so we install cairocffi instead, which
doesn't require manually hacking around with waf. Unfortunately, while
graphite-web HEAD supports cairocffi, the latest released version of
graphite-web does not. This is pretty easily remedied, however, with a simple
patch to graphite/render/glyph.py.

Yes, this is a nasty hack, but until the relevant commit lands in graphite[1]
it's either this or a nasty hack to get pycairo installed into a virtualenv[2].

[1]: https://github.com/graphite-project/graphite-web/commit/d1306e705152c583ca501ac1d4a0fc0d70f6fadd
[2]: http://stackoverflow.com/questions/11491268/install-pycairo-in-virtualenv