Ensure a compiler is present when installing graphite
This is unfortunate, as we'd ideally like to avoid having compilers on
production systems, but it's a tradeoff between cluttering up the system python
libraries with graphite dependencies and having this compiler here.

This is my current preferred option, vs. installing everything except graphite
from apt packages and cluttering up the system (as well as making graphite
susceptible to system package upgrades if you 'pip install anything' on the
system).