Graphite module now takes $version as a parameter
Due to the way that the module was setup beforehand, we'd just install
_any_ version of graphite (carbon, graphite-web) from pip and not
check that we're installing compatible versions.

Due to the addition of the "package==version" string to the end of the
exec calls that we're using, the `$carbon_pip_args_str` and
`$graphite_pip_args_str` variables have been shortened so as not to
cause the PuppetLint tool to show warnings about lines being larger
than 80 lines.