Add beaker acceptance tests
This adds beaker acceptance tests.  There are two possible machine
images. You can run them with the commands:

    BEAKER_set=centos-64-x64 bundle exec rake beaker
    BEAKER_set=ubuntu-1204-x64 bundle exec rake beaker

The machine names correspond to files in
spec/acceptance/nodesets/*.yml. If you don't specify a `BEAKER_set`, it
will look for one called `default`.

I added a basic test to check the default functionality --
carbon-aggregator not running, carbon-cache running and listening on
2003, graphite-web running and listening on 127.0.0.1:8000.

The tests don't pass on centos (as expected at the moment), and I
haven't added them to travis in this commit.