Manual merge of GDS Verify changes to GDS Ops
GDS Verify use a fork of the graphite module. It would be beter of to
use the GDS Operations puppet module and this commit is the first
attempt to add the specific features used in the GDS Verify module.

The features namely are:
 - Ability to specify the user used to run the processes
 - Install from Debian package

To manage user/group permission this commit uses Exec instead of
File resource, simply because the graphite directory can grow very large,
and managing large directories with Puppet can lead to memory starvation.
The exec will use xargs and parralise chmod, so even for large directories
it should run pretty quickly.