Fix pip installation path
Added options for pip to ensure that python packages are installed into
the correct location. The specific option is install_args and without
this the Graphite related packages are installed outside the virtualenv
and can not be found when trying to start the services. By default the
Pip packages are installed in /opt/graphite. Refer to
https://graphite.readthedocs.org/en/latest/install-source.html for more
information.

By using install_args we can pass install options to pip and correctly
install the Python package to root_dir