Support use of Debian/Ubuntu packages
The default behaviour of this module is to install graphite (and
related services) with Python Pip. This commit enables support of using
Debian/Ubuntu packages in addtion to Python Pip.

In theory this should also work with other distros as it uses Puppet
Package resource (and package names can be specified). However this has
not be tested.