restart apps when upgrading packages
Currently, if you change the requested version to cause an upgrade (or
downgrade), the existing service will just keep running.  If this module
wants to manage the installation of the graphite packages, doesn't it
make sense that it should restart the services as well if the package
version changes?

I tried digging in the commit history to see if there was a rationale
for the current behaviour: d83763115 introduced this behaviour, saying
only that it is the "prefered pattern".