Improve acceptance test for graphite-web service
Test that Graphite can successfully serve a request for the dashboard rather
than just testing that it has bound to the port.

This exposes a problem introduced by the `python::pip` changes first in
ed5642d when used with the default `root_dir` of `/opt/graphite`. It
prevents `graphite-web` from being able to load `whisper` and causes it to
respond with a backtrace ending in:

    ImportError: No module named whisper

This was first noticed when bringing up our `graphite-1.management` machine
and recorded in Pivotal [#85067926]. I'm not sure what the fix for this is
yet, so that will follow in a subsequent commit.

PS: I could have use the `curl_on` helper instead of `shell` but it's not
clear to me whether I need to wrap it in an `expect{}` or what value it
provides.