Fix graphite-web service with default root_dir
Add `lib/` to `PYTHONPATH` so that graphite-web is able to load the
`whisper` module when it's installed to the default `root_dir` of
`/opt/graphite`.

This was fixed for custom `root_dir` values in ed5642d but broke the default
use case. I don't really understand the specifics of why, but this seems
like a reasonable fix.