Create metadata.json file
Use of the Modulefile to declare dependencies and other miscellaneous
metadata has been deprecated as of Puppet 3.6. Puppet Labs are soon to
release 4.0, and have been doing a spot of spring cleaning to remove modules
which won't work going forward due to invalid metadata.

Creates a metadata.json file, per the official spec -
https://docs.puppetlabs.com/puppet/latest/reference/modules_publishing.html
- and removes the existing Modulefile.