Move single quote in carbon-cache pre-start
This quote was added in 8a8dc2f1e6aee431ab328a196d408a57fabb9f9c.

I assume it was added to prevent spaces in `root_dir` from removing
multiple files, but it looks weird to have a quote in the middle
of a path:

```
'/opt/graphite'/storage/carbon-cache-a.pid
```

I think moving the quote to the end of the line will have the same
effect but looks a little bit more normal.